@@ -13,20 +13,20 @@
     ]
   ],
   "_from": "etag@>=1.8.0 <1.9.0",
-  "_id": "etag@1.8.0",
+  "_id": "etag@1.8.1",
   "_inCache": true,
   "_installable": true,
   "_location": "/express/etag",
-  "_nodeVersion": "4.7.3",
+  "_nodeVersion": "6.11.1",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/etag-1.8.0.tgz_1487475735517_0.6724899658001959"
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/etag-1.8.1.tgz_1505270623443_0.24458415526896715"
   },
   "_npmUser": {
     "email": "doug@somethingdoug.com",
     "name": "dougwilson"
   },
-  "_npmVersion": "2.15.11",
+  "_npmVersion": "3.10.10",
   "_phantomChildren": {},
   "_requested": {
     "name": "etag",
@@ -40,8 +40,8 @@
     "/express",
     "/express/send"
   ],
-  "_resolved": "https://registry.npmjs.org/etag/-/etag-1.8.0.tgz",
-  "_shasum": "6f631aef336d6c46362b51764044ce216be3c051",
+  "_resolved": "https://registry.npmjs.org/etag/-/etag-1.8.1.tgz",
+  "_shasum": "41ae2eeb65efa62268aebfea83ac7d79299b0887",
   "_shrinkwrap": null,
   "_spec": "etag@~1.8.0",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/express",
@@ -62,20 +62,23 @@
   "description": "Create simple HTTP ETags",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
-    "benchmark": "2.1.3",
-    "eslint": "3.15.0",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-markdown": "1.0.0-beta.3",
-    "eslint-plugin-promise": "3.4.2",
-    "eslint-plugin-standard": "2.0.1",
+    "benchmark": "2.1.4",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "5.1.1",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
-    "seedrandom": "2.4.2"
+    "safe-buffer": "5.1.1",
+    "seedrandom": "2.4.3"
   },
   "directories": {},
   "dist": {
-    "shasum": "6f631aef336d6c46362b51764044ce216be3c051",
-    "tarball": "https://registry.npmjs.org/etag/-/etag-1.8.0.tgz"
+    "shasum": "41ae2eeb65efa62268aebfea83ac7d79299b0887",
+    "tarball": "https://registry.npmjs.org/etag/-/etag-1.8.1.tgz"
   },
   "engines": {
     "node": ">= 0.6"
@@ -86,7 +89,7 @@
     "README.md",
     "index.js"
   ],
-  "gitHead": "16979f788efa8c793c8d07543b4d6aef3d2bfff8",
+  "gitHead": "9b1e3e41df31cda4080833c187120b91a7ce8327",
   "homepage": "https://github.com/jshttp/etag#readme",
   "keywords": [
     "etag",
@@ -114,5 +117,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
   },
-  "version": "1.8.0"
+  "version": "1.8.1"
 }
