@@ -3,95 +3,96 @@
     [
       {
         "name": "internal-ip",
-        "raw": "internal-ip@^1.2.0",
-        "rawSpec": "^1.2.0",
+        "raw": "internal-ip@^2.0.2",
+        "rawSpec": "^2.0.2",
         "scope": null,
-        "spec": ">=1.2.0 <2.0.0",
+        "spec": ">=2.0.2 <3.0.0",
         "type": "range"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack-dev-server"
     ]
   ],
-  "_from": "internal-ip@>=1.2.0 <2.0.0",
-  "_id": "internal-ip@1.2.0",
+  "_from": "internal-ip@>=2.0.2 <3.0.0",
+  "_id": "internal-ip@2.0.2",
   "_inCache": true,
   "_installable": true,
   "_location": "/internal-ip",
-  "_nodeVersion": "4.3.0",
+  "_nodeVersion": "8.4.0",
   "_npmOperationalInternal": {
-    "host": "packages-5-east.internal.npmjs.com",
-    "tmp": "tmp/internal-ip-1.2.0.tgz_1456342188905_0.8255082112737"
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/internal-ip-2.0.2.tgz_1503506543596_0.5833584431093186"
   },
   "_npmUser": {
-    "email": "sindresorhus@gmail.com",
-    "name": "sindresorhus"
+    "email": "npm@silverwind.io",
+    "name": "silverwind"
   },
-  "_npmVersion": "2.14.12",
+  "_npmVersion": "5.3.0",
   "_phantomChildren": {},
   "_requested": {
     "name": "internal-ip",
-    "raw": "internal-ip@^1.2.0",
-    "rawSpec": "^1.2.0",
+    "raw": "internal-ip@^2.0.2",
+    "rawSpec": "^2.0.2",
     "scope": null,
-    "spec": ">=1.2.0 <2.0.0",
+    "spec": ">=2.0.2 <3.0.0",
     "type": "range"
   },
   "_requiredBy": [
     "/webpack-dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/internal-ip/-/internal-ip-1.2.0.tgz",
-  "_shasum": "ae9fbf93b984878785d50a8de1b356956058cf5c",
+  "_resolved": "https://registry.npmjs.org/internal-ip/-/internal-ip-2.0.2.tgz",
+  "_shasum": "bed2b35491e8b42aee087de7614e870908ee80f2",
   "_shrinkwrap": null,
-  "_spec": "internal-ip@^1.2.0",
+  "_spec": "internal-ip@^2.0.2",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack-dev-server",
   "author": {
     "email": "sindresorhus@gmail.com",
     "name": "Sindre Sorhus",
     "url": "sindresorhus.com"
   },
-  "bin": {
-    "internal-ip": "cli.js"
-  },
   "bugs": {
     "url": "https://github.com/sindresorhus/internal-ip/issues"
   },
   "dependencies": {
-    "meow": "^3.3.0"
+    "default-gateway": "^2.0.2",
+    "ipaddr.js": "^1.5.1"
   },
-  "description": "Get your internal IPv4 or IPv6 address",
+  "description": "Get your internal IP address",
   "devDependencies": {
     "ava": "*",
-    "is-ip": "^1.0.0",
     "xo": "*"
   },
   "directories": {},
   "dist": {
-    "shasum": "ae9fbf93b984878785d50a8de1b356956058cf5c",
-    "tarball": "https://registry.npmjs.org/internal-ip/-/internal-ip-1.2.0.tgz"
+    "integrity": "sha512-1OBR5z3Te0FrUtpsPP04XQ2Tsgj+AwujBioVtM9WVW/Jyoe7xR3BIfWnnmg1eUwRuYFelIJM9mQwB7u3bLXD7Q==",
+    "shasum": "bed2b35491e8b42aee087de7614e870908ee80f2",
+    "tarball": "https://registry.npmjs.org/internal-ip/-/internal-ip-2.0.2.tgz"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=6"
   },
   "files": [
-    "index.js",
-    "cli.js"
+    "index.js"
   ],
-  "gitHead": "ef43d204ee3e36dbc0bc8333b5ab816208553269",
-  "homepage": "https://github.com/sindresorhus/internal-ip",
+  "gitHead": "6b17cd6379f400eaae7041b6d5a0e67967828067",
+  "homepage": "https://github.com/sindresorhus/internal-ip#readme",
   "keywords": [
-    "cli-app",
-    "cli",
-    "bin",
     "ip",
-    "ipv4",
     "ipv6",
+    "ipv4",
     "address",
     "internal",
     "local",
-    "machine"
+    "machine",
+    "system",
+    "net",
+    "gateway"
   ],
   "license": "MIT",
   "maintainers": [
+    {
+      "email": "npm@silverwind.io",
+      "name": "silverwind"
+    },
     {
       "email": "sindresorhus@gmail.com",
       "name": "sindresorhus"
@@ -107,5 +108,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "1.2.0"
+  "version": "2.0.2"
 }
