@@ -3,46 +3,46 @@
     [
       {
         "name": "opn",
-        "raw": "opn@4.0.2",
-        "rawSpec": "4.0.2",
+        "raw": "opn@^5.1.0",
+        "rawSpec": "^5.1.0",
         "scope": null,
-        "spec": "4.0.2",
-        "type": "version"
+        "spec": ">=5.1.0 <6.0.0",
+        "type": "range"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack-dev-server"
     ]
   ],
-  "_from": "opn@4.0.2",
-  "_id": "opn@4.0.2",
+  "_from": "opn@>=5.1.0 <6.0.0",
+  "_id": "opn@5.1.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/opn",
-  "_nodeVersion": "4.4.2",
+  "_nodeVersion": "8.0.0",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/opn-4.0.2.tgz_1463477356148_0.1645404922310263"
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/opn-5.1.0.tgz_1497093567063_0.9740056861191988"
   },
   "_npmUser": {
     "email": "sindresorhus@gmail.com",
     "name": "sindresorhus"
   },
-  "_npmVersion": "3.9.0",
+  "_npmVersion": "5.0.0",
   "_phantomChildren": {},
   "_requested": {
     "name": "opn",
-    "raw": "opn@4.0.2",
-    "rawSpec": "4.0.2",
+    "raw": "opn@^5.1.0",
+    "rawSpec": "^5.1.0",
     "scope": null,
-    "spec": "4.0.2",
-    "type": "version"
+    "spec": ">=5.1.0 <6.0.0",
+    "type": "range"
   },
   "_requiredBy": [
     "/webpack-dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/opn/-/opn-4.0.2.tgz",
-  "_shasum": "7abc22e644dff63b0a96d5ab7f2790c0f01abc95",
+  "_resolved": "https://registry.npmjs.org/opn/-/opn-5.1.0.tgz",
+  "_shasum": "72ce2306a17dbea58ff1041853352b4a8fc77519",
   "_shrinkwrap": null,
-  "_spec": "opn@4.0.2",
+  "_spec": "opn@^5.1.0",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack-dev-server",
   "author": {
     "email": "sindresorhus@gmail.com",
@@ -53,8 +53,7 @@
     "url": "https://github.com/sindresorhus/opn/issues"
   },
   "dependencies": {
-    "object-assign": "^4.0.1",
-    "pinkie-promise": "^2.0.0"
+    "is-wsl": "^1.1.0"
   },
   "description": "A better node-open. Opens stuff like websites, files, executables. Cross-platform.",
   "devDependencies": {
@@ -63,17 +62,18 @@
   },
   "directories": {},
   "dist": {
-    "shasum": "7abc22e644dff63b0a96d5ab7f2790c0f01abc95",
-    "tarball": "https://registry.npmjs.org/opn/-/opn-4.0.2.tgz"
+    "integrity": "sha512-iPNl7SyM8L30Rm1sjGdLLheyHVw5YXVfi3SKWJzBI7efxRwHojfRFjwE/OLM6qp9xJYMgab8WicTU1cPoY+Hpg==",
+    "shasum": "72ce2306a17dbea58ff1041853352b4a8fc77519",
+    "tarball": "https://registry.npmjs.org/opn/-/opn-5.1.0.tgz"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js",
     "xdg-open"
   ],
-  "gitHead": "b56b0e981ee377d3b04c57a4e6748ad2793ada17",
+  "gitHead": "ebae31e9117ae3ca5c9ee642bafa9449f1e237c2",
   "homepage": "https://github.com/sindresorhus/opn#readme",
   "keywords": [
     "app",
@@ -119,5 +119,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "4.0.2"
+  "version": "5.1.0"
 }
