@@ -3,46 +3,46 @@
     [
       {
         "name": "supports-color",
-        "raw": "supports-color@^3.1.1",
-        "rawSpec": "^3.1.1",
+        "raw": "supports-color@^4.2.1",
+        "rawSpec": "^4.2.1",
         "scope": null,
-        "spec": ">=3.1.1 <4.0.0",
+        "spec": ">=4.2.1 <5.0.0",
         "type": "range"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack-dev-server"
     ]
   ],
-  "_from": "supports-color@>=3.1.1 <4.0.0",
-  "_id": "supports-color@3.2.3",
+  "_from": "supports-color@>=4.2.1 <5.0.0",
+  "_id": "supports-color@4.4.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/webpack-dev-server/supports-color",
-  "_nodeVersion": "7.2.0",
+  "_nodeVersion": "8.2.1",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/supports-color-3.2.3.tgz_1484526472497_0.7697830176912248"
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/supports-color-4.4.0.tgz_1504162477210_0.36436482798308134"
   },
   "_npmUser": {
-    "email": "i.am.qix@gmail.com",
-    "name": "qix"
+    "email": "sindresorhus@gmail.com",
+    "name": "sindresorhus"
   },
-  "_npmVersion": "3.10.9",
+  "_npmVersion": "5.3.0",
   "_phantomChildren": {},
   "_requested": {
     "name": "supports-color",
-    "raw": "supports-color@^3.1.1",
-    "rawSpec": "^3.1.1",
+    "raw": "supports-color@^4.2.1",
+    "rawSpec": "^4.2.1",
     "scope": null,
-    "spec": ">=3.1.1 <4.0.0",
+    "spec": ">=4.2.1 <5.0.0",
     "type": "range"
   },
   "_requiredBy": [
     "/webpack-dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/supports-color/-/supports-color-3.2.3.tgz",
-  "_shasum": "65ac0504b3954171d8a64946b2ae3cbb8a5f54f6",
+  "_resolved": "https://registry.npmjs.org/supports-color/-/supports-color-4.4.0.tgz",
+  "_shasum": "883f7ddabc165142b2a61427f3352ded195d1a3e",
   "_shrinkwrap": null,
-  "_spec": "supports-color@^3.1.1",
+  "_spec": "supports-color@^4.2.1",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/webpack-dev-server",
   "author": {
     "email": "sindresorhus@gmail.com",
@@ -54,27 +54,28 @@
     "url": "https://github.com/chalk/supports-color/issues"
   },
   "dependencies": {
-    "has-flag": "^1.0.0"
+    "has-flag": "^2.0.0"
   },
   "description": "Detect whether a terminal supports color",
   "devDependencies": {
-    "mocha": "*",
-    "require-uncached": "^1.0.2",
+    "ava": "*",
+    "import-fresh": "^2.0.0",
     "xo": "*"
   },
   "directories": {},
   "dist": {
-    "shasum": "65ac0504b3954171d8a64946b2ae3cbb8a5f54f6",
-    "tarball": "https://registry.npmjs.org/supports-color/-/supports-color-3.2.3.tgz"
+    "integrity": "sha512-rKC3+DyXWgK0ZLKwmRsrkyHVZAjNkfzeehuFWdGGcqGDTZFH73+RH6S/RDAAxl9GusSjZSUWYLmT9N5pzXFOXQ==",
+    "shasum": "883f7ddabc165142b2a61427f3352ded195d1a3e",
+    "tarball": "https://registry.npmjs.org/supports-color/-/supports-color-4.4.0.tgz"
   },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">=4"
   },
   "files": [
     "index.js",
     "browser.js"
   ],
-  "gitHead": "c2394c1f77a0f64c4b2d0db33a3df3b841e6fc15",
+  "gitHead": "d2e32f77030ddf583b95b79bc3f6417241472980",
   "homepage": "https://github.com/chalk/supports-color#readme",
   "keywords": [
     "color",
@@ -96,8 +97,7 @@
     "capability",
     "detect",
     "truecolor",
-    "16m",
-    "million"
+    "16m"
   ],
   "license": "MIT",
   "maintainers": [
@@ -118,14 +118,7 @@
     "url": "git+https://github.com/chalk/supports-color.git"
   },
   "scripts": {
-    "test": "xo && mocha",
-    "travis": "mocha"
+    "test": "xo && ava"
   },
-  "version": "3.2.3",
-  "xo": {
-    "envs": [
-      "node",
-      "mocha"
-    ]
-  }
+  "version": "4.4.0"
 }
