@@ -90,47 +90,51 @@
     "express": "^4.13.3",
     "html-entities": "^1.2.0",
     "http-proxy-middleware": "~0.17.4",
-    "internal-ip": "^1.2.0",
+    "internal-ip": "^2.0.2",
     "ip": "^1.1.5",
     "loglevel": "^1.4.1",
-    "opn": "4.0.2",
+    "opn": "^5.1.0",
     "portfinder": "^1.0.9",
     "selfsigned": "^1.9.1",
     "serve-index": "^1.7.2",
     "sockjs": "0.3.18",
     "sockjs-client": "1.1.4",
     "spdy": "^3.4.1",
-    "strip-ansi": "^3.0.0",
-    "supports-color": "^3.1.1",
+    "strip-ansi": "^3.0.1",
+    "supports-color": "^4.2.1",
     "webpack-dev-middleware": "^1.11.0",
-    "yargs": "^6.0.0"
+    "yargs": "^6.6.0"
   },
   "description": "Serves a webpack app. Updates the browser on changes.",
   "devDependencies": {
     "codecov.io": "^0.1.6",
-    "css-loader": "~0.26.1",
-    "eslint": "^3.4.0",
-    "file-loader": "~0.10.0",
+    "css-loader": "^0.28.5",
+    "eslint": "^4.5.0",
+    "eslint-config-webpack": "^1.2.5",
+    "eslint-plugin-import": "^2.7.0",
+    "file-loader": "^0.11.2",
     "istanbul": "^0.4.5",
-    "jquery": "^2.2.0",
+    "jquery": "^3.2.1",
     "less": "^2.5.1",
-    "less-loader": "~2.2.0",
+    "less-loader": "^4.0.5",
     "mocha": "^3.0.2",
-    "mocha-sinon": "^1.1.6",
+    "mocha-sinon": "^2.0.0",
     "pug": "^2.0.0-beta5",
     "pug-loader": "^2.3.0",
-    "should": "^11.1.0",
-    "sinon": "^1.17.6",
-    "style-loader": "~0.13.0",
-    "supertest": "^2.0.1",
+    "should": "^12.0.0",
+    "sinon": "^3.2.1",
+    "style-loader": "^0.18.2",
+    "supertest": "^3.0.0",
+    "uglifyjs-webpack-plugin": "^1.0.0-beta.2",
     "url-loader": "~0.5.6",
     "webpack": "^3.0.0",
-    "ws": "^1.1.1"
+    "ws": "^3.1.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "21580f5a08cd065c71144cf6f61c345bca59a8b8",
-    "tarball": "https://registry.npmjs.org/webpack-dev-server/-/webpack-dev-server-2.7.1.tgz"
+    "integrity": "sha512-wD9bs+Z1uwvf3Jc+8ZkyMI0Xi+aJJYjC2UZplOWoo/vStelK5Mv62X2uXYEYIQEjy9wJQMzC0fEFqQsg7vVEIg==",
+    "shasum": "abd61f410778cc4c843d7cebbf41465b1ab7734c",
+    "tarball": "https://registry.npmjs.org/webpack-dev-server/-/webpack-dev-server-2.8.2.tgz"
   },
   "engines": {
     "node": ">=4.7"
