@@ -13,20 +13,20 @@
     ]
   ],
   "_from": "websocket-driver@>=0.5.1",
-  "_id": "websocket-driver@0.6.5",
+  "_id": "websocket-driver@0.7.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/websocket-driver",
-  "_nodeVersion": "4.4.4",
+  "_nodeVersion": "4.8.4",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/websocket-driver-0.6.5.tgz_1463730072239_0.9899731166660786"
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/websocket-driver-0.7.0.tgz_1505163256798_0.4708031218033284"
   },
   "_npmUser": {
     "email": "jcoglan@gmail.com",
     "name": "jcoglan"
   },
-  "_npmVersion": "2.15.1",
+  "_npmVersion": "2.15.11",
   "_phantomChildren": {},
   "_requested": {
     "name": "websocket-driver",
@@ -40,8 +40,8 @@
     "/faye-websocket",
     "/sockjs-client/faye-websocket"
   ],
-  "_resolved": "https://registry.npmjs.org/websocket-driver/-/websocket-driver-0.6.5.tgz",
-  "_shasum": "5cb2556ceb85f4373c6d8238aa691c8454e13a36",
+  "_resolved": "https://registry.npmjs.org/websocket-driver/-/websocket-driver-0.7.0.tgz",
+  "_shasum": "0caf9d2d755d93aee049d4bdd0d3fe2cca2a24eb",
   "_shrinkwrap": null,
   "_spec": "websocket-driver@>=0.5.1",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/faye-websocket",
@@ -54,22 +54,26 @@
     "url": "https://github.com/faye/websocket-driver-node/issues"
   },
   "dependencies": {
+    "http-parser-js": ">=0.4.0",
     "websocket-extensions": ">=0.1.1"
   },
   "description": "WebSocket protocol handler with pluggable I/O",
   "devDependencies": {
-    "jstest": "",
-    "permessage-deflate": ""
+    "jstest": "*",
+    "permessage-deflate": "*"
   },
   "directories": {},
   "dist": {
-    "shasum": "5cb2556ceb85f4373c6d8238aa691c8454e13a36",
-    "tarball": "https://registry.npmjs.org/websocket-driver/-/websocket-driver-0.6.5.tgz"
+    "shasum": "0caf9d2d755d93aee049d4bdd0d3fe2cca2a24eb",
+    "tarball": "https://registry.npmjs.org/websocket-driver/-/websocket-driver-0.7.0.tgz"
   },
   "engines": {
-    "node": ">=0.6.0"
+    "node": ">=0.8.0"
   },
-  "gitHead": "c4494ff88ac196f726bbb77a301c2177124b199e",
+  "files": [
+    "lib"
+  ],
+  "gitHead": "8128bec1e03c4eeb3c3bdb250a4356a4c69fe10f",
   "homepage": "https://github.com/faye/websocket-driver-node",
   "keywords": [
     "websocket"
@@ -92,5 +96,5 @@
   "scripts": {
     "test": "jstest spec/runner.js"
   },
-  "version": "0.6.5"
+  "version": "0.7.0"
 }
