@@ -13,16 +13,20 @@
     ]
   ],
   "_from": "websocket-extensions@>=0.1.1",
-  "_id": "websocket-extensions@0.1.1",
+  "_id": "websocket-extensions@0.1.2",
   "_inCache": true,
   "_installable": true,
   "_location": "/websocket-extensions",
-  "_nodeVersion": "0.12.0",
+  "_nodeVersion": "4.8.4",
+  "_npmOperationalInternal": {
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/websocket-extensions-0.1.2.tgz_1505062237525_0.6339729737956077"
+  },
   "_npmUser": {
     "email": "jcoglan@gmail.com",
     "name": "jcoglan"
   },
-  "_npmVersion": "2.5.1",
+  "_npmVersion": "2.15.11",
   "_phantomChildren": {},
   "_requested": {
     "name": "websocket-extensions",
@@ -35,8 +39,8 @@
   "_requiredBy": [
     "/websocket-driver"
   ],
-  "_resolved": "https://registry.npmjs.org/websocket-extensions/-/websocket-extensions-0.1.1.tgz",
-  "_shasum": "76899499c184b6ef754377c2dbb0cd6cb55d29e7",
+  "_resolved": "https://registry.npmjs.org/websocket-extensions/-/websocket-extensions-0.1.2.tgz",
+  "_shasum": "0e18781de629a18308ce1481650f67ffa2693a5d",
   "_shrinkwrap": null,
   "_spec": "websocket-extensions@>=0.1.1",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/websocket-driver",
@@ -51,17 +55,20 @@
   "dependencies": {},
   "description": "Generic extension manager for WebSocket connections",
   "devDependencies": {
-    "jstest": ""
+    "jstest": "*"
   },
   "directories": {},
   "dist": {
-    "shasum": "76899499c184b6ef754377c2dbb0cd6cb55d29e7",
-    "tarball": "https://registry.npmjs.org/websocket-extensions/-/websocket-extensions-0.1.1.tgz"
+    "shasum": "0e18781de629a18308ce1481650f67ffa2693a5d",
+    "tarball": "https://registry.npmjs.org/websocket-extensions/-/websocket-extensions-0.1.2.tgz"
   },
   "engines": {
-    "node": ">=0.6.0"
+    "node": ">=0.8.0"
   },
-  "gitHead": "89104ddd48ccbbff20237e3fc5c123b4d4c5d9c1",
+  "files": [
+    "lib"
+  ],
+  "gitHead": "5f040a15af1343f512af2fce9b12339044ce25eb",
   "homepage": "http://github.com/faye/websocket-extensions-node",
   "keywords": [
     "websocket"
@@ -84,5 +91,5 @@
   "scripts": {
     "test": "jstest spec/runner.js"
   },
-  "version": "0.1.1"
+  "version": "0.1.2"
 }
