@@ -4,8 +4,18 @@
  * MIT Licensed
  */
 
+'use strict'
+
+/**
+ * Simple expression to split token list.
+ * @private
+ */
+
+var TOKEN_LIST_REGEXP = / *, */
+
 /**
  * Module exports.
+ * @public
  */
 
 module.exports = forwarded
@@ -13,18 +23,20 @@ module.exports = forwarded
 /**
  * Get all addresses in the request, using the `X-Forwarded-For` header.
  *
- * @param {Object} req
- * @api public
+ * @param {object} req
+ * @return {array}
+ * @public
  */
 
-function forwarded(req) {
+function forwarded (req) {
   if (!req) {
     throw new TypeError('argument req is required')
   }
 
   // simple header parsing
   var proxyAddrs = (req.headers['x-forwarded-for'] || '')
-    .split(/ *, */)
+    .trim()
+    .split(TOKEN_LIST_REGEXP)
     .filter(Boolean)
     .reverse()
   var socketAddr = req.connection.remoteAddress
