@@ -1,41 +1,36 @@
 'use strict';
-var os = require('os');
+const os = require('os');
+const defaultGateway = require('default-gateway');
+const ipaddr = require('ipaddr.js');
 
-var type = {
-	v4: {
-		def: '127.0.0.1',
-		family: 'IPv4'
-	},
-	v6: {
-		def: '::1',
-		family: 'IPv6'
-	}
+const defaults = {
+	v6: '::1',
+	v4: '127.0.0.1'
 };
 
-function internalIp(version) {
-	var options = type[version];
-	var ret = options.def;
-	var interfaces = os.networkInterfaces();
+function internalIp(family) {
+	return defaultGateway[family]().then(result => {
+		const interfaces = os.networkInterfaces();
+		const gatewayIp = ipaddr.parse(result.gateway);
+		let ret;
 
-	Object.keys(interfaces).forEach(function (el) {
-		interfaces[el].forEach(function (el2) {
-			if (!el2.internal && el2.family === options.family) {
-				ret = el2.address;
-			}
+		// Look for the matching interface in all local interfaces
+		Object.keys(interfaces).some(name => {
+			return interfaces[name].some(addr => {
+				const prefix = ipaddr.parse(addr.netmask).prefixLengthFromSubnetMask();
+				const net = ipaddr.parseCIDR(`${addr.address}/${prefix}`);
+
+				if (net[0].kind() === gatewayIp.kind() && gatewayIp.match(net)) {
+					ret = net[0].toString();
+				}
+
+				return Boolean(ret);
+			});
 		});
-	});
 
-	return ret;
+		return ret ? ret : defaults[family];
+	}).catch(() => defaults[family]);
 }
 
-function v4() {
-	return internalIp('v4');
-}
-
-function v6() {
-	return internalIp('v6');
-}
-
-module.exports = v4;
-module.exports.v4 = v4;
-module.exports.v6 = v6;
+module.exports.v6 = () => internalIp('v6');
+module.exports.v4 = () => internalIp('v4');
