@@ -78,11 +78,12 @@ var Async = function (_Component) {
 		_this._cache = props.cache === defaultCache ? {} : props.cache;
 
 		_this.state = {
+			inputValue: '',
 			isLoading: false,
 			options: props.options
 		};
 
-		_this._onInputChange = _this._onInputChange.bind(_this);
+		_this.onInputChange = _this.onInputChange.bind(_this);
 		return _this;
 	}
 
@@ -106,9 +107,9 @@ var Async = function (_Component) {
 			}
 		}
 	}, {
-		key: 'clearOptions',
-		value: function clearOptions() {
-			this.setState({ options: [] });
+		key: 'componentWillUnmount',
+		value: function componentWillUnmount() {
+			this._callback = null;
 		}
 	}, {
 		key: 'loadOptions',
@@ -119,7 +120,7 @@ var Async = function (_Component) {
 
 			var cache = this._cache;
 
-			if (cache && cache.hasOwnProperty(inputValue)) {
+			if (cache && Object.prototype.hasOwnProperty.call(cache, inputValue)) {
 				this.setState({
 					options: cache[inputValue]
 				});
@@ -163,8 +164,8 @@ var Async = function (_Component) {
 			}
 		}
 	}, {
-		key: '_onInputChange',
-		value: function _onInputChange(inputValue) {
+		key: 'onInputChange',
+		value: function onInputChange(inputValue) {
 			var _props = this.props,
 			    ignoreAccents = _props.ignoreAccents,
 			    ignoreCase = _props.ignoreCase,
@@ -184,18 +185,11 @@ var Async = function (_Component) {
 				onInputChange(transformedInputValue);
 			}
 
+			this.setState({ inputValue: inputValue });
 			this.loadOptions(transformedInputValue);
 
 			// Return the original input value to avoid modifying the user's view of the input while typing.
 			return inputValue;
-		}
-	}, {
-		key: 'inputValue',
-		value: function inputValue() {
-			if (this.select) {
-				return this.select.state.inputValue;
-			}
-			return '';
 		}
 	}, {
 		key: 'noResultsText',
@@ -204,10 +198,10 @@ var Async = function (_Component) {
 			    loadingPlaceholder = _props2.loadingPlaceholder,
 			    noResultsText = _props2.noResultsText,
 			    searchPromptText = _props2.searchPromptText;
-			var isLoading = this.state.isLoading;
+			var _state = this.state,
+			    inputValue = _state.inputValue,
+			    isLoading = _state.isLoading;
 
-
-			var inputValue = this.inputValue();
 
 			if (isLoading) {
 				return loadingPlaceholder;
@@ -230,10 +224,13 @@ var Async = function (_Component) {
 			var _props3 = this.props,
 			    children = _props3.children,
 			    loadingPlaceholder = _props3.loadingPlaceholder,
-			    placeholder = _props3.placeholder;
-			var _state = this.state,
-			    isLoading = _state.isLoading,
-			    options = _state.options;
+			    multi = _props3.multi,
+			    onChange = _props3.onChange,
+			    placeholder = _props3.placeholder,
+			    value = _props3.value;
+			var _state2 = this.state,
+			    isLoading = _state2.isLoading,
+			    options = _state2.options;
 
 
 			var props = {
@@ -242,18 +239,12 @@ var Async = function (_Component) {
 				options: isLoading && loadingPlaceholder ? [] : options,
 				ref: function ref(_ref) {
 					return _this3.select = _ref;
-				},
-				onChange: function onChange(newValues) {
-					if (_this3.props.multi && _this3.props.value && newValues.length > _this3.props.value.length) {
-						_this3.clearOptions();
-					}
-					_this3.props.onChange(newValues);
 				}
 			};
 
 			return children(_extends({}, this.props, props, {
 				isLoading: isLoading,
-				onInputChange: this._onInputChange
+				onInputChange: this.onInputChange
 			}));
 		}
 	}]);
