@@ -6,14 +6,9 @@ Object.defineProperty(exports, "__esModule", {
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
+var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
+
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; /*!
-                                                                                                                                                                                                                                                                                Copyright (c) 2017 Jed Watson.
-                                                                                                                                                                                                                                                                                Licensed under the MIT License (MIT), see
-                                                                                                                                                                                                                                                                                http://jedwatson.github.io/react-select
-                                                                                                                                                                                                                                                                              */
-
 
 var _react = require('react');
 
@@ -69,20 +64,16 @@ function _classCallCheck(instance, Constructor) { if (!(instance instanceof Cons
 
 function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; } /*!
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Copyright (c) 2017 Jed Watson.
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Licensed under the MIT License (MIT), see
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 http://jedwatson.github.io/react-select
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               */
 
-function stringifyValue(value) {
-	var valueType = typeof value === 'undefined' ? 'undefined' : _typeof(value);
-	if (valueType === 'string') {
-		return value;
-	} else if (valueType === 'object') {
-		return JSON.stringify(value);
-	} else if (valueType === 'number' || valueType === 'boolean') {
-		return String(value);
-	} else {
-		return '';
-	}
-}
+
+var stringifyValue = function stringifyValue(value) {
+	return typeof value === 'string' ? value : value !== null && JSON.stringify(value) || '';
+};
 
 var stringOrNode = _propTypes2.default.oneOfType([_propTypes2.default.string, _propTypes2.default.node]);
 
@@ -96,28 +87,9 @@ var Select = function (_React$Component) {
 
 		var _this = _possibleConstructorReturn(this, (Select.__proto__ || Object.getPrototypeOf(Select)).call(this, props));
 
-		_this.handleTouchOutside = _this.handleTouchOutside.bind(_this);
-		_this.handleTouchMove = _this.handleTouchMove.bind(_this);
-		_this.handleTouchStart = _this.handleTouchStart.bind(_this);
-		_this.handleTouchEnd = _this.handleTouchEnd.bind(_this);
-		_this.handleTouchEndClearValue = _this.handleTouchEndClearValue.bind(_this);
-		_this.handleMouseDown = _this.handleMouseDown.bind(_this);
-		_this.handleMouseDownOnArrow = _this.handleMouseDownOnArrow.bind(_this);
-		_this.handleMouseDownOnMenu = _this.handleMouseDownOnMenu.bind(_this);
-		_this.handleInputFocus = _this.handleInputFocus.bind(_this);
-		_this.handleInputBlur = _this.handleInputBlur.bind(_this);
-		_this.handleInputChange = _this.handleInputChange.bind(_this);
-		_this.handleInputValueChange = _this.handleInputValueChange.bind(_this);
-		_this.handleKeyDown = _this.handleKeyDown.bind(_this);
-		_this.handleValueClick = _this.handleValueClick.bind(_this);
-		_this.handleMenuScroll = _this.handleMenuScroll.bind(_this);
-		_this.handleRequired = _this.handleRequired.bind(_this);
-		_this.getOptionLabel = _this.getOptionLabel.bind(_this);
-		_this.onOptionRef = _this.onOptionRef.bind(_this);
-		_this.clearValue = _this.clearValue.bind(_this);
-		_this.removeValue = _this.removeValue.bind(_this);
-		_this.selectValue = _this.selectValue.bind(_this);
-		_this.focusOption = _this.focusOption.bind(_this);
+		['clearValue', 'focusOption', 'handleInputBlur', 'handleInputChange', 'handleInputFocus', 'handleInputValueChange', 'handleKeyDown', 'handleMenuScroll', 'handleMouseDown', 'handleMouseDownOnArrow', 'handleMouseDownOnMenu', 'handleRequired', 'handleTouchOutside', 'handleTouchMove', 'handleTouchStart', 'handleTouchEnd', 'handleTouchEndClearValue', 'handleValueClick', 'getOptionLabel', 'onOptionRef', 'removeValue', 'selectValue'].forEach(function (fn) {
+			return _this[fn] = _this[fn].bind(_this);
+		});
 
 		_this.state = {
 			inputValue: '',
@@ -190,8 +162,10 @@ var Select = function (_React$Component) {
 				var menuDOM = _reactDom2.default.findDOMNode(this.menu);
 				var focusedRect = focusedDOM.getBoundingClientRect();
 				var menuRect = menuDOM.getBoundingClientRect();
-				if (focusedRect.bottom > menuRect.bottom || focusedRect.top < menuRect.top) {
+				if (focusedRect.bottom > menuRect.bottom) {
 					menuDOM.scrollTop = focusedDOM.offsetTop + focusedDOM.clientHeight - menuDOM.offsetHeight;
+				} else if (focusedRect.top < menuRect.top) {
+					menuDOM.scrollTop = focusedDOM.offsetTop;
 				}
 			}
 			if (this.props.scrollMenuIntoView && this.menuContainer) {
@@ -331,7 +305,7 @@ var Select = function (_React$Component) {
 				});
 			} else {
 				// otherwise, focus the input and open the menu
-				this._openAfterFocus = this.props.openAfterFocus;
+				this._openAfterFocus = this.props.openOnClick;
 				this.focus();
 			}
 		}
@@ -544,7 +518,7 @@ var Select = function (_React$Component) {
 			if (!this.props.onMenuScrollToBottom) return;
 			var target = event.target;
 
-			if (target.scrollHeight > target.offsetHeight && !(target.scrollHeight - target.offsetHeight - target.scrollTop)) {
+			if (target.scrollHeight > target.offsetHeight && target.scrollHeight - target.offsetHeight - target.scrollTop <= 0) {
 				this.props.onMenuScrollToBottom();
 			}
 		}
@@ -617,37 +591,42 @@ var Select = function (_React$Component) {
 			if (this.props.autoBlur) {
 				this.blurInput();
 			}
-			if (!this.props.onChange) return;
 			if (this.props.required) {
 				var required = this.handleRequired(value, this.props.multi);
 				this.setState({ required: required });
 			}
-			if (this.props.simpleValue && value) {
-				value = this.props.multi ? value.map(function (i) {
-					return i[_this3.props.valueKey];
-				}).join(this.props.delimiter) : value[this.props.valueKey];
+			if (this.props.onChange) {
+				if (this.props.simpleValue && value) {
+					value = this.props.multi ? value.map(function (i) {
+						return i[_this3.props.valueKey];
+					}).join(this.props.delimiter) : value[this.props.valueKey];
+				}
+				this.props.onChange(value);
 			}
-			this.props.onChange(value);
 		}
 	}, {
 		key: 'selectValue',
 		value: function selectValue(value) {
 			var _this4 = this;
 
-			//NOTE: update value in the callback to make sure the input value is empty so that there are no styling issues (Chrome had issue otherwise)
-			this.hasScrolledToOption = false;
+			// NOTE: we actually add/set the value in a callback to make sure the
+			// input value is empty to avoid styling issues in Chrome
+			if (this.props.closeOnSelect) {
+				this.hasScrolledToOption = false;
+			}
 			if (this.props.multi) {
 				var updatedValue = this.props.onSelectResetsInput ? '' : this.state.inputValue;
 				this.setState({
+					focusedIndex: null,
 					inputValue: this.handleInputValueChange(updatedValue),
-					focusedIndex: null
+					isOpen: !this.props.closeOnSelect
 				}, function () {
 					_this4.addValue(value);
 				});
 			} else {
 				this.setState({
-					isOpen: false,
 					inputValue: this.handleInputValueChange(''),
+					isOpen: !this.props.closeOnSelect,
 					isPseudoFocused: this.state.isFocused
 				}, function () {
 					_this4.setValue(value);
@@ -903,8 +882,7 @@ var Select = function (_React$Component) {
 
 			var ariaOwns = (0, _classnames2.default)((_classNames = {}, _defineProperty(_classNames, this._instancePrefix + '-list', isOpen), _defineProperty(_classNames, this._instancePrefix + '-backspace-remove-message', this.props.multi && !this.props.disabled && this.state.isFocused && !this.state.inputValue), _classNames));
 
-			// TODO: Check how this project includes Object.assign()
-			var inputProps = Object.assign({}, this.props.inputProps, {
+			var inputProps = _extends({}, this.props.inputProps, {
 				role: 'combobox',
 				'aria-expanded': '' + isOpen,
 				'aria-owns': ariaOwns,
@@ -964,7 +942,6 @@ var Select = function (_React$Component) {
 	}, {
 		key: 'renderClear',
 		value: function renderClear() {
-
 			if (!this.props.clearable || this.props.value === undefined || this.props.value === null || this.props.multi && !this.props.value.length || this.props.disabled || this.props.isLoading) return;
 			var clear = this.props.clearRenderer();
 
@@ -1234,6 +1211,7 @@ Select.propTypes = {
 	clearRenderer: _propTypes2.default.func, // create clearable x element
 	clearValueText: stringOrNode, // title for the "clear" control
 	clearable: _propTypes2.default.bool, // should it be possible to reset value
+	closeOnSelect: _propTypes2.default.bool, // whether to close the menu when a value is selected
 	deleteRemoves: _propTypes2.default.bool, // whether backspace removes an item if there is no text input
 	delimiter: _propTypes2.default.string, // delimiter to use to join multiple values for the hidden field value
 	disabled: _propTypes2.default.bool, // whether the Select is disabled or not
@@ -1269,7 +1247,7 @@ Select.propTypes = {
 	onOpen: _propTypes2.default.func, // fires when the menu is opened
 	onSelectResetsInput: _propTypes2.default.bool, // whether input is cleared on select (works only for multiselect)
 	onValueClick: _propTypes2.default.func, // onClick handler for value labels: function (value, event) {}
-	openAfterFocus: _propTypes2.default.bool, // boolean to enable opening dropdown when focused
+	openOnClick: _propTypes2.default.bool, // boolean to control opening the menu when the control is clicked
 	openOnFocus: _propTypes2.default.bool, // always open options menu on focus
 	optionClassName: _propTypes2.default.string, // additional class(es) to apply to the <Option /> elements
 	optionComponent: _propTypes2.default.func, // option component to render in dropdown
@@ -1302,6 +1280,7 @@ Select.defaultProps = {
 	clearAllText: 'Clear all',
 	clearRenderer: _defaultClearRenderer2.default,
 	clearValueText: 'Clear value',
+	closeOnSelect: true,
 	deleteRemoves: true,
 	delimiter: ',',
 	disabled: false,
@@ -1322,6 +1301,7 @@ Select.defaultProps = {
 	onBlurResetsInput: true,
 	onSelectResetsInput: true,
 	onCloseResetsInput: true,
+	openOnClick: true,
 	optionComponent: _Option2.default,
 	pageSize: 5,
 	placeholder: 'Select...',
