@@ -3,7 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
 	value: true
 });
-exports.Creatable = exports.AsyncCreatable = exports.Async = undefined;
+exports.Value = exports.Creatable = exports.AsyncCreatable = exports.Async = undefined;
 
 var _Select = require('./Select');
 
@@ -21,13 +21,19 @@ var _Creatable = require('./Creatable');
 
 var _Creatable2 = _interopRequireDefault(_Creatable);
 
+var _Value = require('./Value');
+
+var _Value2 = _interopRequireDefault(_Value);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 _Select2.default.Async = _Async2.default;
 _Select2.default.AsyncCreatable = _AsyncCreatable2.default;
 _Select2.default.Creatable = _Creatable2.default;
+_Select2.default.Value = _Value2.default;
 
 exports.default = _Select2.default;
 exports.Async = _Async2.default;
 exports.AsyncCreatable = _AsyncCreatable2.default;
 exports.Creatable = _Creatable2.default;
+exports.Value = _Value2.default;
