@@ -52,11 +52,12 @@ export default class Async extends Component {
 		this._cache = props.cache === defaultCache ? {} : props.cache;
 
 		this.state = {
+			inputValue: '',
 			isLoading: false,
 			options: props.options,
 		};
 
-		this._onInputChange = this._onInputChange.bind(this);
+		this.onInputChange = this.onInputChange.bind(this);
 	}
 
 	componentDidMount () {
@@ -75,8 +76,8 @@ export default class Async extends Component {
 		}
 	}
 
-	clearOptions() {
-		this.setState({ options: [] });
+	componentWillUnmount () {
+		this._callback = null;
 	}
 
 	loadOptions (inputValue) {
@@ -85,7 +86,7 @@ export default class Async extends Component {
 
 		if (
 			cache &&
-			cache.hasOwnProperty(inputValue)
+			Object.prototype.hasOwnProperty.call(cache, inputValue)
 		) {
 			this.setState({
 				options: cache[inputValue]
@@ -132,7 +133,7 @@ export default class Async extends Component {
 		}
 	}
 
-	_onInputChange (inputValue) {
+	onInputChange (inputValue) {
 		const { ignoreAccents, ignoreCase, onInputChange } = this.props;
 		let transformedInputValue = inputValue;
 
@@ -148,24 +149,16 @@ export default class Async extends Component {
 			onInputChange(transformedInputValue);
 		}
 
+		this.setState({ inputValue });
 		this.loadOptions(transformedInputValue);
 
 		// Return the original input value to avoid modifying the user's view of the input while typing.
 		return inputValue;
 	}
 
-	inputValue() {
-		if (this.select) {
-			return this.select.state.inputValue;
-		}
-		return '';
-	}
-
 	noResultsText() {
 		const { loadingPlaceholder, noResultsText, searchPromptText } = this.props;
-		const { isLoading } = this.state;
-
-		const inputValue = this.inputValue();
+		const { inputValue, isLoading } = this.state;
 
 		if (isLoading) {
 			return loadingPlaceholder;
@@ -181,7 +174,7 @@ export default class Async extends Component {
 	}
 
 	render () {
-		const { children, loadingPlaceholder, placeholder } = this.props;
+		const { children, loadingPlaceholder, multi, onChange, placeholder, value } = this.props;
 		const { isLoading, options } = this.state;
 
 		const props = {
@@ -189,19 +182,13 @@ export default class Async extends Component {
 			placeholder: isLoading ? loadingPlaceholder : placeholder,
 			options: (isLoading && loadingPlaceholder) ? [] : options,
 			ref: (ref) => (this.select = ref),
-			onChange: (newValues) => {
-				if (this.props.multi && this.props.value && (newValues.length > this.props.value.length)) {
-					this.clearOptions();
-				}
-				this.props.onChange(newValues);
-			}
 		};
 
 		return children({
 			...this.props,
 			...props,
 			isLoading,
-			onInputChange: this._onInputChange
+			onInputChange: this.onInputChange
 		});
 	}
 }
