@@ -2,14 +2,17 @@ import Select from './Select';
 import Async from './Async';
 import AsyncCreatable from './AsyncCreatable';
 import Creatable from './Creatable';
+import Value from './Value';
 
 Select.Async = Async;
 Select.AsyncCreatable = AsyncCreatable;
 Select.Creatable = Creatable;
+Select.Value = Value;
 
 export default Select;
 export {
 	Async,
 	AsyncCreatable,
-	Creatable
+	Creatable,
+	Value
 };
