@@ -1958,7 +1958,8 @@ describe('Select', () => {
 				value: '',
 				options: options,
 				searchable: false,
-				multi: true
+				multi: true,
+				closeOnSelect: false,
 			}, {
 				wireUpOnChangeToValue: true
 			});
@@ -2961,7 +2962,7 @@ describe('Select', () => {
 
 				instance = createControl({
 					options: defaultOptions,
-					onFocus: onFocus
+					onFocus: onFocus,
 				});
 			});
 
@@ -2971,27 +2972,51 @@ describe('Select', () => {
 			});
 		});
 
-		describe('openAfterFocus', () => {
-
-			var openAfterFocus;
-
-			beforeEach(() => {
-				openAfterFocus = sinon.spy();
-
+		describe('openOnClick', () => {
+			it('should open the menu on click when true', () => {
 				instance = createControl({
 					options: defaultOptions,
-					openAfterFocus: true
+					openOnClick: true,
+				}, {
+					initialFocus: false,
 				});
+				TestUtils.Simulate.mouseDown(ReactDOM.findDOMNode(instance).querySelector('.Select-control'), { button: 0 });
+				// need to simulate the focus on the input, it does not happen in jsdom
+				findAndFocusInputControl();
+				expect(instance.state.isOpen, 'to be true');
 			});
 
-			it('should show the options when focused', () => {
-				instance.focus();
+			it('should not open the menu on click when false', () => {
+				instance = createControl({
+					options: defaultOptions,
+					openOnClick: false,
+				}, {
+					initialFocus: false,
+				});
+				TestUtils.Simulate.mouseDown(ReactDOM.findDOMNode(instance).querySelector('.Select-control'), { button: 0 });
+				// need to simulate the focus on the input, it does not happen in jsdom
+				findAndFocusInputControl();
+				expect(instance.state.isOpen, 'to be falsy');
+			});
+		});
 
-				if (instance.state.isFocused && instance.state.openAfterFocus) {
-					expect(instance.state.isOpen, 'to equal', true);
-				}
+		describe('openOnFocus', () => {
+			// Note: createControl automatically focuses the control
+			it('should open the menu on focus when true', () => {
+				instance = createControl({
+					options: defaultOptions,
+					openOnFocus: true,
+				});
+				expect(instance.state.isOpen, 'to be true');
 			});
 
+			it('should open the menu on focus when false', () => {
+				instance = createControl({
+					options: defaultOptions,
+					openOnFocus: false,
+				});
+				expect(instance.state.isOpen, 'to be falsy');
+			});
 		});
 
 		describe('onValueClick', () => {
@@ -3864,7 +3889,8 @@ describe('Select', () => {
 						{ value: 'five', label: 'label five' }
 					],
 					value: [ 'three', 'two' ],
-					multi: true
+					multi: true,
+					closeOnSelect: false,
 				}, {
 					wireUpOnChangeToValue: true
 				});
