@@ -1,41 +1,42 @@
-var SockJS = require("sockjs-client");
+'use strict';
 
-var retries = 0;
-var sock = null;
+const SockJS = require('sockjs-client');
+
+let retries = 0;
+let sock = null;
 
 function socket(url, handlers) {
-	sock = new SockJS(url);
+  sock = new SockJS(url);
 
-	sock.onopen = function() {
-		retries = 0;
-	}
+  sock.onopen = function onopen() {
+    retries = 0;
+  };
 
-	sock.onclose = function() {
-		if(retries === 0)
-			handlers.close();
+  sock.onclose = function onclose() {
+    if (retries === 0) { handlers.close(); }
 
-		// Try to reconnect.
-		sock = null;
+    // Try to reconnect.
+    sock = null;
 
-		// After 10 retries stop trying, to prevent logspam.
-		if(retries <= 10) {
-			// Exponentially increase timeout to reconnect.
-			// Respectfully copied from the package `got`.
-			var retryInMs = 1000 * Math.pow(2, retries) + Math.random() * 100;
-			retries += 1;
+    // After 10 retries stop trying, to prevent logspam.
+    if (retries <= 10) {
+      // Exponentially increase timeout to reconnect.
+      // Respectfully copied from the package `got`.
+      // eslint-disable-next-line no-mixed-operators, no-restricted-properties
+      const retryInMs = 1000 * Math.pow(2, retries) + Math.random() * 100;
+      retries += 1;
 
-			setTimeout(function() {
-				socket(url, handlers);
-			}, retryInMs);
-		}
-	};
+      setTimeout(function cb() {
+        socket(url, handlers);
+      }, retryInMs);
+    }
+  };
 
-	sock.onmessage = function(e) {
-		// This assumes that all data sent via the websocket is JSON.
-		var msg = JSON.parse(e.data);
-		if(handlers[msg.type])
-			handlers[msg.type](msg.data);
-	};
+  sock.onmessage = function onmessage(e) {
+    // This assumes that all data sent via the websocket is JSON.
+    const msg = JSON.parse(e.data);
+    if (handlers[msg.type]) { handlers[msg.type](msg.data); }
+  };
 }
 
 module.exports = socket;
